Updated the repos and refactored code
