+++ GroupProject/src/main/java/Bookstore/repositories/BookstoreRepository.java
@@ -2,9 +2,8 @@ package Bookstore.repositories;
 
 import org.springframework.data.repository.PagingAndSortingRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
-import team.GroupProject.Bookstore;
 
 @RepositoryRestResource(collectionResourceRel = "bookstores", path = "bookstores")
-public interface BookstoreRepository extends PagingAndSortingRepository<Bookstore, Long> {
+public interface BookstoreRepository extends PagingAndSortingRepository<Bookstore, Long>{
     Bookstore findById(long id);
 }
