+++ GroupProject/src/main/java/Bookstore/repositories/CartRepository.java
@@ -1,4 +1,10 @@
 package Bookstore.repositories;
 
-public interface CartRepository {
+
+import org.springframework.data.repository.PagingAndSortingRepository;
+import org.springframework.data.rest.core.annotation.RepositoryRestResource;
+
+@RepositoryRestResource(collectionResourceRel = "carts", path = "carts")
+public interface CartRepository extends PagingAndSortingRepository<Cart,Long>{
+    Cart findById(long id);
 }
