+++ GroupProject/src/main/java/Bookstore/repositories/OwnerRepository.java
@@ -1,12 +1,12 @@
 package Bookstore.repositories;
 
+import org.springframework.data.repository.PagingAndSortingRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
-import team.GroupProject.Owner;
 
 import java.util.List;
 
 @RepositoryRestResource(collectionResourceRel = "owners", path = "owners")
-public interface OwnerRepository {
+public interface OwnerRepository extends PagingAndSortingRepository<Owner,Long> {
     Owner findById(long id);
     List<Owner> findByName(String name);
 }
