+++ GroupProject/src/main/java/Bookstore/repositories/UserRepository.java
@@ -1,15 +1,15 @@
 package Bookstore.repositories;
 
+
 import org.springframework.data.repository.PagingAndSortingRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 import org.springframework.stereotype.Repository;
-import team.GroupProject.User;
 
 import java.util.List;
 
 @Repository
 @RepositoryRestResource(collectionResourceRel = "users", path = "users")
-public interface UserRepository extends PagingAndSortingRepository<User,Long> {
+public interface UserRepository extends PagingAndSortingRepository<User,Long>{
 
     User findById(long id);
