+++ GroupProject/src/main/java/team/GroupProject/Customer.java
@@ -1,5 +1,7 @@
 package team.GroupProject;
 
+import Bookstore.repositories.Cart;
+
 import java.util.Set;
 
 public class Customer {
